Swap skills and experience component names
