+++ src/App.js
@@ -8,7 +8,7 @@ registerIcons();
 function App() {
   const Home = lazy(() => import('./components/sections/Home/Home'));
   const About = lazy(() => import('./components/sections/About/About'));
-  const Skills = lazy(() => import('./components/sections/Skills/Skills'));
+  const Experience = lazy(() => import('./components/sections/Experience/Experience'));
 
   // @TODO: Add proper fallback when the site is loading with complete styling 
   return (
@@ -16,7 +16,7 @@ function App() {
       <Suspense fallback={<div>Loading...</div>}>
         <Home />
         <About />
-        <Skills />
+        <Experience />
       </Suspense> 
     </div>
   );
