+++ src/components/common/molecules/Skills/Skills.module.scss
@@ -1,10 +1,10 @@
-.experience {
+.skills {
     :global {
         .tech-content {
             @include responsive('margin-bottom', '30px');
 
             .row {
-                margin-bottom: 15px;
+                @include responsive('margin-bottom', '15px', '20px', '20px', '20px', '20px');
             }
 
             .skill {
@@ -23,19 +23,18 @@
 
                 &__logos {
                     flex-flow: row wrap;
-                    gap: 2.5em 1em;
                     flex: 1 1 auto;
-                    justify-content: space-evenly;
+                    @include responsive('justify-content', 'space-evenly', 'center', 'center', 'center', 'center');
+                    @include responsive('gap', '2.5em 1em', '2.5em .25em', '2.5em .25em', '2.5em .25em', '2.5em .25em');
 
                     .logo-col {
-                        // padding-bottom: 0;
-                        // height: 100px;
+                        @include responsive('align-items', 'center', 'center', 'center', 'center', 'center');
 
                         .logo {
                             @include responsive('width', '100%', '55px', '55px', '55px', '55px');
+                            @include responsive('height', 'auto', 'fit-content', 'fit-content', 'fit-content', 'fit-content');
                             max-width: 100%;
                             background-size: cover;
-                            height: auto;
                             display: block;
                             margin: 0 auto;
                             transition: all .3s ease-in-out;
