+++ src/components/sections/Experience/Experience.module.scss
@@ -1,4 +1,4 @@
-.skills {
+.experience {
     color: #313131;
     padding-top: 2.5rem;
     padding-bottom: 2.5rem;
