+++ src/components/sections/Experience/Experience.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Skills from './Skills';
+import Experience from './Experience';
 import { render, cleanup, waitFor } from '@testing-library/react';
 import { mockAllIsIntersecting } from 'react-intersection-observer/test-utils'
 
@@ -10,7 +10,7 @@ let component;
 beforeEach(() => {
   component = (
     <React.Suspense fallback="loading">
-      <Skills />
+      <Experience />
     </React.Suspense>
   );
 });
@@ -18,7 +18,7 @@ beforeEach(() => {
 test('should render properly', async () => {
   const { getByText } = render(component);
 
-  const lazyElement = await waitFor(() => getByText(/Skills/i));
+  const lazyElement = await waitFor(() => getByText(/Experience/i));
 
   expect(lazyElement).toBeInTheDocument();
 });
@@ -28,7 +28,7 @@ test('should add active class to title when the component is in view', async() =
 
   mockAllIsIntersecting(true);
 
-  const lazyElement = await waitFor(() => getByText(/Skills/i));
+  const lazyElement = await waitFor(() => getByText(/Experience/i));
 
   expect(lazyElement.closest('div.underline')).toBeInTheDocument();
   expect(lazyElement.closest('div.underline').classList.contains('underline-active')).toBe(true);
